+++ lib/widget/app_body.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:lux/widget/proxy_list_card.dart';
+import 'package:url_launcher/url_launcher.dart';
 import 'package:window_manager/window_manager.dart';
 
 import '../core/core_config.dart';
@@ -8,12 +9,14 @@ import '../core/core_manager.dart';
 class AppBody extends StatefulWidget {
   final CoreManager coreManager;
   final String curProxyInfo;
+  final String dashboardUrl;
   final void Function(String) onCurProxyInfoChange;
   const AppBody(
       {super.key,
       required this.coreManager,
       required this.curProxyInfo,
-      required this.onCurProxyInfoChange});
+      required this.onCurProxyInfoChange,
+      required this.dashboardUrl});
 
   @override
   State<AppBody> createState() => _AppBodyState();
@@ -26,10 +29,20 @@ class _AppBodyState extends State<AppBody> with WindowListener {
 
   bool isLoadingProxyRadio = false;
 
+  var isCollapsedMap = <String, bool>{};
+
   Future<void> refreshProxyList() async {
     final value = await widget.coreManager.getProxyList();
     setState(() {
       proxyListGroup = value;
+      for (var group in proxyListGroup.groups) {
+        var key = group.url;
+        if (!isCollapsedMap.containsKey(key)) {
+          setState(() {
+            isCollapsedMap[key] = true;
+          });
+        }
+      }
     });
   }
 
@@ -82,6 +95,28 @@ class _AppBodyState extends State<AppBody> with WindowListener {
     windowManager.removeListener(this);
   }
 
+  bool getIsCollapsed(ProxyList item) {
+    return isCollapsedMap.containsKey(item.url)
+        ? (isCollapsedMap[item.url] as bool)
+        : true;
+  }
+
+  void handleCollapse(ProxyList item) {
+    setState(() {
+      isCollapsedMap[item.url] = !getIsCollapsed(item);
+    });
+  }
+
+  void _handleDeleteItem(ProxyItem item) async {
+    await widget.coreManager.deleteProxies([item.id]);
+    await refreshData();
+  }
+
+  void _handleEditItem(ProxyItem item) async {
+    final editingUrl = "${widget.dashboardUrl}&mode=edit&proxyId=${item.id}";
+    launchUrl(Uri.parse(editingUrl));
+  }
+
   @override
   Widget build(BuildContext context) {
     return RadioGroup<String>(
@@ -95,6 +130,11 @@ class _AppBodyState extends State<AppBody> with WindowListener {
                   return ProxyListCard(
                     proxyList: proxyListGroup.groups[index],
                     key: Key(proxyListGroup.groups[index].url),
+                    isCollapsed: getIsCollapsed(proxyListGroup.groups[index]),
+                    onCollapse: () =>
+                        {handleCollapse(proxyListGroup.groups[index])},
+                    onDeleteItem: _handleDeleteItem,
+                    onEditItem: _handleEditItem,
                   );
                 },
               ));