+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ### New features ✨
 
-feat: category setting items
+* feat: support editing and deleting proxy item from native UI
+* feat: show proxy type in native UI
+* feat: handle core starting error
+* feat: tip for bottom bar
 
 ### Bug fixes 🐛
 
-* fix(windows): missing restarting app alert modal when the mode is changed to Tun
-* fix: fail to request subscription url without user agent
+* fix: collapse state of card is lost when scrolling
 
 ### Other changes